Se cambia el nombre del proyecto, carpetas y app
