+++ app/build.gradle
@@ -20,7 +20,7 @@ android {
     buildToolsVersion '30.0.2'
 
     defaultConfig {
-        applicationId "com.example.evidenciafinal"
+        applicationId "com.mafervicas.fitTracking"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
