+++ app/src/androidTest/java/com/mafervicas/fitTracking/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.evidenciafinal;
+package com.mafervicas.fitTracking;
 
 import android.content.Context;
