+++ app/src/main/java/com/mafervicas/fitTracking/VPAdapter.java
@@ -1,4 +1,4 @@
-package com.example.evidenciafinal;
+package com.mafervicas.fitTracking;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
