+++ app/src/main/java/com/mafervicas/fitTracking/views/Dashboard.java
@@ -1,4 +1,4 @@
-package com.example.evidenciafinal.views;
+package com.mafervicas.fitTracking.views;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatButton;
@@ -12,7 +12,7 @@ import android.widget.ImageButton;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.evidenciafinal.R;
+import com.mafervicas.fitTracking.R;
 
 import java.text.DecimalFormat;
