+++ app/src/main/java/com/mafervicas/fitTracking/views/Registration.java
@@ -1,4 +1,4 @@
-package com.example.evidenciafinal.views;
+package com.mafervicas.fitTracking.views;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -9,13 +9,12 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.Button;
-import android.widget.ImageButton;
 import android.widget.ProgressBar;
 import android.widget.Spinner;
 import android.widget.Toast;
 
-import com.example.evidenciafinal.R;
-import com.example.evidenciafinal.model.DatosDB;
+import com.mafervicas.fitTracking.R;
+import com.mafervicas.fitTracking.model.DatosDB;
 import com.google.android.material.textfield.TextInputEditText;
 
 import java.text.DateFormat;
@@ -23,17 +22,18 @@ import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
 
-public class NewInformation extends AppCompatActivity {
-
+public class Registration extends AppCompatActivity {
     //Starting with the initialization of variables
     private Button btnSiguiente;
     DatosDB myDb;
-    TextInputEditText tiEdad, tiAltura, tiPeso;
+    TextInputEditText tiNombre, tiEdad, tiAltura, tiPeso;
     SharedPreferences sp;
     ProgressBar myProgressBar;
     Spinner s;
-    ImageButton buttonReturnDashboard;
 
+    //String with the information of the water & Exercise
+    private String stringQueMandaremos;
+    public static final String STRING_QUE_MANDAREMOS = "com.example.evidenciafinal.STRING_QUE_MANDAREMOS";
     String exerciseFreq;
     Integer exerciseMultiply;
 
@@ -46,13 +46,14 @@ public class NewInformation extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_new_information);
+        setContentView(R.layout.activity_registration);
 
         //Connect XML Components
+        tiNombre = (TextInputEditText) findViewById(R.id.tiNombre);
         tiPeso = (TextInputEditText) findViewById(R.id.tiPeso);
         tiAltura = (TextInputEditText) findViewById(R.id.tiAltura);
         tiEdad = (TextInputEditText) findViewById(R.id.tiEdad);
-        btnSiguiente = (Button) findViewById(R.id.btnGuardar);
+        btnSiguiente = (Button) findViewById(R.id.btnSiguiente);
         myProgressBar = (ProgressBar) findViewById(R.id.progressBar);
         s = (Spinner) findViewById(R.id.spinnerExercise);
 
@@ -70,15 +71,6 @@ public class NewInformation extends AppCompatActivity {
             }
         });
 
-        //Button Dashboard
-        buttonReturnDashboard = findViewById(R.id.includedButton).findViewById(R.id.ibLogo);
-        buttonReturnDashboard.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openMainActivity();
-            }
-        });
-
         //Call DB constructor Llama el constructor
         myDb = new DatosDB(this);
 
@@ -90,10 +82,12 @@ public class NewInformation extends AppCompatActivity {
                 String peso = tiPeso.getText().toString();
                 String altura = tiAltura.getText().toString();
                 String edad = tiEdad.getText().toString();
+                String nombre = tiNombre.getText().toString();
+                String nameStr = tiNombre.getText().toString();
 
-                if(peso.isEmpty() || altura.isEmpty() || edad.isEmpty()){
+                if(peso.isEmpty() || altura.isEmpty() || edad.isEmpty() || nombre.isEmpty()){
                     //Missing fields
-                    Toast.makeText(NewInformation.this, "Favor de llenar todos los campos",Toast.LENGTH_LONG).show();
+                    Toast.makeText(Registration.this, "Favor de llenar todos los campos",Toast.LENGTH_LONG).show();
                 } else {
                     //Show ProgressBar
                     myProgressBar.setVisibility(View.VISIBLE);
@@ -118,6 +112,7 @@ public class NewInformation extends AppCompatActivity {
                     //Add name to Shared preferences
                     sp = getSharedPreferences("UserPreferences", Context.MODE_PRIVATE);
                     SharedPreferences.Editor editor = sp.edit();
+                    editor.putString("name", nameStr);
                     editor.putString("imc", IMC.toString());
                     editor.putString("water", Ingesta.toString());
                     editor.putString("kcals", kcalsFinales.toString());
@@ -152,17 +147,13 @@ public class NewInformation extends AppCompatActivity {
         if(addInformation = true) {
             //Make invisible progress bar
             myProgressBar.setVisibility(View.INVISIBLE);
-            Toast.makeText(NewInformation.this, "Agregado correctamente", Toast.LENGTH_SHORT).show();
+            Toast.makeText(Registration.this, "Agregado correctamente", Toast.LENGTH_SHORT).show();
             Intent intent = new Intent(this, Dashboard.class);
+            intent.putExtra(STRING_QUE_MANDAREMOS, stringQueMandaremos);
             startActivity(intent);
 
         } else {
-            Toast.makeText(NewInformation.this, "Problemas al insertar", Toast.LENGTH_LONG).show();
+            Toast.makeText(Registration.this, "Problemas al insertar", Toast.LENGTH_LONG).show();
         }
-    }
-
-    public void openMainActivity(){
-        Intent intent = new Intent(this, Dashboard.class);
-        startActivity(intent);
     }
 }
