+++ app/src/main/res/navigation/nav_graph2.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.evidenciafinal.FirstFragment"
+        android:name="com.mafervicas.fitTracking.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/Second2Fragment"
-        android:name="com.example.evidenciafinal.Second2Fragment"
+        android:name="com.mafervicas.fitTracking.Second2Fragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second2">
