+++ app/src/main/res/navigation/nav_graph3.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.evidenciafinal.FirstFragment"
+        android:name="com.mafervicas.fitTracking.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.example.evidenciafinal.SecondFragment"
+        android:name="com.mafervicas.fitTracking.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
